This fix addresses Issue #335, re: Gloomshade Grove
The issue addressed by Thrumbar in issue #335 explains that some quests
show up with the first objective showing as intended but the remaining
objectives which shouldn't show up are listed as "nil" in the quest
watch window. I was able to resolve this issue in my version by removing
the objective lines and updating the quest giver ids according to
Wowhead.