Edit to "Their Hunt Continues"
This quest wouldn't show up in the database even after a reload. So I
removed the objective lines and added the "Start" and "End" lines which
fixed this issue.

About removing objective lines: While I understand that objective lines
are helpful in some quests, many have multiple unnecessary nil lines
that point to the possible locations of one or more objective targets.
These lines usually have to be reduced or eliminated to allow Carbonite
to actually point to the correct locations. I believe removing many of
these lines on quests that don't need them will reduce the size of the
files and free up processing resources for actual play.