Edited Children's Week quests
I edited the Children's Week quests so they'd show up in the database.

Note: For quest #13956, which ends in Un'Goro Crater, I set the
coordinates for the waygate in Sholazar Basin as directed in the quest
description. I then set the end coordinates in Un'Goro Crater near the
Etymidian which will update the way point when the player summons the
orphan.

I also updated "Back to the Orphanage" quest data for the alternate
database listings that you get if you already have the Curious Oracle
Hatching and/or Curious Wolvar Pup pets.

Finally, for most of the quests, I didn't include the end coordinates
for quests that could only be turned into the orphans since they had to
be summoned to turn them in eliminating the need for coordinate to help
find them.

As usual, I welcome feedback if needed so I can fix or add what I may
have missed.