Edits - Tanaan, Garrison Fishing Shack, and Barn
Fishing Shack:
Added Fishing Shack quests for recruiting Nat Pagle (A/H), one fishing
daily on Alliance side, and the level 2 Barn quest for the Alliance.

Quest #36611 ( [A True Draenor Angler) which is a neutral quest is
bugged (on Bliz side). Regardless of what faction you are on, Blizzard's
map points only to Frostwall to turn in this quest. So I added this
quest twice which actually worked. The only issue I noticed was that the
way point is a little over 200 yards from the actual turn-in location
even though I verified the coordinates while standing in front of Nat
Pagle at the Fishing Shack. I won't know if this will be the same on
other alts. I also need to test this on the Horde side to see if it does
the same thing.

Barn: Quest #36274 (Bigger Trap, Better Rewards" (Alliance only)
Since the second objective of this quest is hidden until the first
objective is compete, I added it replacing "nil" with the quest title so
I could add both possible plot locations to make sure the direction
arrow will (hopefully) point toward the player's Barn when he/she
returns to place the work order.

Tanaan:  Added quest #39513 (Disarming Sha'naar) with the objective line
pointing to the entrance of the area where the quest takes place.

Quest #39574 (Pressing the Attack):
I added objective lines pointing to the 7 (I think I got them all)
locations where players can complete the bonus tasks for this quest. One
note on this though, since the first line points toward the Fel Forge,
that is where the arrow will point when ticking this quest. So all this
addition does is add the quest to the database so it can be ticked.

Final note, I edited the last three quests before #35671 (A Gathering of
Shadows) correcting their turn-in locations so the player finds Reshad
to turn those quests in and so he can lead him/her to Shadow-Sage Iskar
to complete "A Gathering of Shadows".