Tanaan Dailies - both factions
Added:
Assaults on the Fel Forge and Throne of Kil'jaeden
Demonslayer of Tanaan with objective lines pointing to a few mob
locations.
Edited:
Disarming Sha'naar -  corrected faction designation for Horde version.
Unseen Influence - added objective lines pointing to a few places where
Fel Corrupted Apexis fragments may be looted.

It appears that these edits and additions may have eliminated the need
for a reload to get these quests to show up in the database. I would
need to test this further or get verification from someone willing to
test this as well.