More Tanaan Dailies and minor edits.
I added more Tanaan dailies, edited some some pre-TJ quests and re-did
"A True Draenor Angler" which is a bugged quest that always points the
player to Frostwall regardless of his/her faction. This is Blizzard's
bug, so I could find no way around it neither by duplicating the quest
in the database nor by adding two quest endings. So I deleted the
duplicate, designated the quest as neutral and left the objectives in
tact so players will still be able to track the locations for the
Lunkers they have to find.

Also, I duplicated the quests "Unknowable Power" and "Unseen Influence"
assigning the appropriate faction and turn-in coordinates. This was the
only way I could get the objective and turn-in locations to show up
properly in the database. Since the objectives can be completed while
doing the other quests however, the objectives may not be necessary. So
they may be deleted at a later date, but I think leaving the quests
duplicated will ensure players are directed to their proper outposts to
turn them in since Blizzard (imho) messed this quest up by putting the
same quest giver and only one "neutral" quest in two separate locations.