Edits I forgot to push a long time ago
The most recent edit I did on this file was adding quest #36385, "Legacy
of the Ancients" which is the quest you get from a Tome you get when you
cut down a Petrified Ancient. Since it is listed on Wowhead as a level
90 quest, I added it to this file instead of the 91-100 file. Also, the
quest number falls in line with other level 90 WoD quests.

The older quests here are just minor edits I did to get the quests to
show up in the database or update quest giver id numbers.