Merge latest commits from Rythal master to my Legion (#9)
* Made it so that you could add multiple waypoints when clicking on icons

This will allow you to hold "shift" when clicking on icons on the map, such as those added by HandyNotes, so that you can set multiple waypoints on these icons.

* align minimap blocks of Broken Isle

* Fix for quest detail texture and green quest blobs.

I think i got some good fixes here. 

Some tip for us, we HAVE to change all SetTexture with solid colors to SetColorTexture in all Carbonite filles. 

This was change in WOW API 7.0.3:
*The Texture method Texture:SetTexture(R, G, B[, A]) no longer works properly in many cases. Use the new method Texture:SetColorTexture instead.*

* Fixing SetTexture.

* Fixing one line (my falut).

* Another fix for bad replacement.

* green map fixes

* zone adds

* Many SetTexture Updates

* noticed dungeons off by 1 level.. this should fix

* fixes button updates after texture color patch