I added all of the Demon Hunter starting quests...
I added all the Demon Hunter Staring quests starting from the quest
after character creation and ending with Alliance and Horde quests that
send their characters to their Demon Hunter outposts.

Note: The hud waypoint has trouble in the Demon Ward of the Vault of
Wardens. The arrow won't point to the correct quest coordinates even
though the map locations show up correctly. Before that, the hud works
fine. I think this may be because Vault of the Wardens is still set up
as an instance instead of a small area within the zone.

Therefore quests beginning in Vault of Wardens and ending with the
meeting with Archmage Kadgar, may be off by as high as 23 yards
according to the waypoint hud.

Later when I recover from Demon Hunter burnout after running both
factions back to back in two days, I'll run another after a few more
updates have been pushed to see if this changes either with a blizz
patch or a Carbonite update. :-)