Added the last pre-Legion quests
Added:
44182 - "City Under Siege" Even though this is a neutral quest, I had to
add it twice since the portal to Dalaran is different for the Horde and
Alliance.
44421 - "Anamalous Anomalies"
44500 - "Author! Author!

Note: Unlike the quest "The Council of the Six", the quest you get from
Kadgar's Upgraded Minion for "City Under Siege" immediatly flags as
"complete" so the objective pointing to the portals in Orgrimmar and
Stormwind will be skipped, and the way point will point to Deadwind
Pass. I don't know if this can be fixed in Carbonite to override this,
so I left the objectives for the portals in anyway just in case.