Edited "City Under Siege"
Rather listing quest #44182 "City Under Siege" twice which required a
reload to show up, I deleted one record, set the quest as neutral, and
put both quest objective locations pointing to the portals in Stormwind
Keep and Cleft of Shadow on the same objective line.

With this fix, the quest shows up as intended, but still auto completes
sending the player directly to DW Pass instead of pointing first to the
portals. I may play with this a little more later, but for now this is
the best I can do.