Cleanup of Watch Window and Log
Full Cleanup of watch window, removing not in database objectives.  Added code to allow carbonite log to show the objectives, and not error out when a person clicked on an objective not in the carbonite database.  My next project is to find and update database from blizzard data, and make the questing more dynamic.