Cleanup of code
Reading all that gave me a pain in the head.  Noticed some were caps in some places and lower case in others.  As a most likely gesture I straightened out the code and made it in line with the way all data pulls from Blizzard was presented.  Fixed spaces so all are tab's and not some tab some space bar and some both..